requirements-backend.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
dataclasses==0.8
# Flask==2.0.3
idna==3.3
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.28
ordereddict==1.1
persiantools==2.2.0
protobuf==3.19.4
pytz==2021.3
requests==2.27.1
six==1.16.0
typing-extensions==4.1.1
urllib3==1.26.8
# uWSGI==2.0.20
Werkzeug==2.0.3
zipp==3.6.0
